ContentView renamed to MainView Child views moved as fragmented views within MainView for better code readability Total scoreview addedd based on a scoreCard flag in the MainView
